Merge Schedule MS with ResourcePool MS and start work on the ResourcePool and Faculty classes and Repos

See merge request cse2115/2022-2023/SEM14a!8

+++ nodes-microservice/src/main/java/nl/tudelft/sem/template/nodes/domain/node/NodeRepository.java
@@ -5,10 +5,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 /**
- * A DDD repository for quering and persisting node aggregate roots.
+ * A DDD repository for querying and persisting node aggregate roots.
  */
 @Repository
-public interface NodeRepository extends JpaRepository<Node, Integer> {
+public interface NodeRepository extends JpaRepository<Node, Long> {
 
     Optional<Node> findByName(Name name);
